Structure for tests, moved compression away from util
