+++ src/codeu/chat/compression/CompressionEngine.java
@@ -1,14 +1,15 @@
-package codeu.chat.util;
-
-import codeu.chat.common.Message;
-
-import com.diogoduailibe.lzstring4j.LZString;
-
 /**
  * @author Eric Zhuang, CodeU Project Group 6
  * @description Converts a message into a compressed string format in order
  * to minimize bandwidth when sending messages
 */
+
+package codeu.chat.compression;
+
+import codeu.chat.common.Message;
+
+import com.diogoduailibe.lzstring4j.LZString;
+
 public final class CompressionEngine {
 
     /**
