+++ test/codeu/chat/TestRunner.java
@@ -29,6 +29,7 @@ public final class TestRunner {
              codeu.chat.server.BasicControllerTest.class,
              codeu.chat.server.RawControllerTest.class,
              codeu.chat.util.store.StoreTest.class,
+             codeu.chat.compression.CompressionTest.class,
              codeu.chat.database.DatabaseTest.class
          );
       for (final Failure failure : result.getFailures()) {
